### 2023-11-01 #6009 @editorialbot
[PRE REVIEW]: xeofs: Comprehensive EOF analysis in Python with xarray
**Submitting author:** @nicrie (<a href="http://orcid.org/0000-0003-3357-1742">Niclas Rieger</a>)
**Repository:** https://github.com/nicrie/xeofs
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v.1.2.2
