### 2023-11-01 #6009 @kthyng
Hi @nicrie and thanks for your submission. I saw the note with your submission. Was there a software review as part of the publication below?
N. Rieger, Á. Corral, E. Olmedo, and A. Turiel, “Lagged Teleconnections of Climate Variables Identified via Complex Rotated Maximum Covariance Analysis,” Journal of Climate, vol. 34, no. 24, pp. 9861–9878, Dec. 2021, doi: 10.1175/JCLI-D-21-0244.1.
