### 2023-11-02 #6009 @nicrie
Hey @kthyng, no there wasn't. We just included it with the paper so others could check our work if they wanted to. It wasn't a must-have for the journal, just our way of being open about our research.
