### 2023-11-03 f35cc85
feat: serialize model to DataSet
