### 2023-11-04 7ae7245
docs: add missing DOIs
