+++ publication/paper.bib
@@ -99,6 +99,7 @@ Publisher: Ubiquity Press},
 @article{guilloteau_rotated_2020,
 	title = {Rotated spectral principal component analysis ({rsPCA}) for identifying dynamical modes of variability in climate systems},
 	url = {http://arxiv.org/abs/2004.11411},
+	doi = {10.1175/JCLI-D-20-0266.1},
 	abstract = {Spectral {PCA} ({sPCA}), in contrast to classical {PCA}, offers the advantage of identifying organized spatiotemporal patterns within speciﬁc frequency bands and extracting dynamical modes. However, the unavoidable tradeoff between frequency resolution and robustness of the {PCs} leads to high sensitivity to noise and overﬁtting, which limits the interpretation of the {sPCA} results. We propose herein a simple non-parametric implementation of the {sPCA} using the continuous analytic Morlet wavelet as a robust estimator of the cross-spectral matrices with good frequency resolution. To improve the interpretability of the results when several modes of similar amplitude exist within the same frequency band, we propose a rotation of eigenvectors that optimizes the spatial smoothness in the phase domain. The developed method, called rotated spectral {PCA} ({rsPCA}), is tested on synthetic data simulating propagating waves and shows impressive performance even with high levels of noise in the data. Applied to historical sea surface temperature ({SST}) time series over the Paciﬁc Ocean, the method accurately captures the El Nin˜o-Southern Oscillation ({ENSO}) at low frequency (2 to 7 years periodicity). At high frequencies (sub-annual periodicity), at which several extratropical patterns of similar amplitude are identiﬁed, the {rsPCA} successfully unmixes the underlying modes, revealing spatially coherent patterns with robust propagation dynamics. Identiﬁcation of higher frequency space-time climate modes holds promise for seasonal to subseasonal prediction and for diagnostic analysis of climate models.},
 	journaltitle = {{arXiv}:2004.11411 [physics]},
 	author = {Guilloteau, Clément and Mamalakis, Antonios and Vulis, Lawrence and Georgiou, Tryphon T. and Foufoula-Georgiou, Efi},
@@ -134,6 +135,7 @@ Publisher: Ubiquity Press},
 @book{hannachi_patterns_2021,
 	title = {Patterns Identification and Data Mining in Weather and Climate},
 	isbn = {978-3-030-67072-6},
+	doi = {10.1007/978-3-030-67073-3},
 	series = {Springer Atmospheric Sciences},
 	pagetotal = {600},
 	publisher = {Springer International Publishing},
@@ -159,6 +161,7 @@ Publisher: Ubiquity Press},
 
 @article{vinod_canonical_1976,
 	title = {Canonical ridge and econometrics of joint production},
+	doi = {10.1016/0304-4076(76)90010-5},
 	volume = {4},
 	pages = {147--166},
 	number = {2},
@@ -172,6 +175,7 @@ Publisher: Ubiquity Press},
 
 @article{bretherton_intercomparison_1992,
 	title = {An intercomparison of methods for finding coupled patterns in climate data},
+	doi = {10.1175/1520-0442(1992)005<0541:AIOMFF>2.0.CO;2},
 	volume = {5},
 	pages = {541--560},
 	number = {6},
@@ -310,6 +314,7 @@ Section: Monthly Weather Review},
 
 @article{rasmusson_biennial_1981,
 	title = {Biennial variations in surface temperature over the United States as revealed by singular decomposition},
+	doi = {10.1175/1520-0493(1981)109<0587:BVISTO>2.0.CO;2},
 	volume = {109},
 	pages = {587--598},
 	number = {3},
