### 2023-11-05 b18c3e0
refactor: Initialize Preprocessor upfront
Previously, the Preprocessor was fully initialized within the fit method
,as the decision between using a DataArrayStacker or a DataSetStacker
depended on the data being fitted. This commit consolidates both
stackers into a unified Stacker class, enabling the complete
initialization of the Preprocessor at the start of the model
initialization process.
