### 2023-11-07 bf9de1b
feat: full serialization with zarr and datatree
