+++ xeofs/preprocessing/transformer.py
@@ -1,9 +1,15 @@
+from __future__ import annotations
 from abc import ABC
-from typing import Optional
+from typing import Optional, Dict, TYPE_CHECKING
 from typing_extensions import Self
 from abc import abstractmethod
 
+import pandas as pd
+import xarray as xr
 from sklearn.base import BaseEstimator, TransformerMixin
+
+if TYPE_CHECKING:
+    from datatree import DataTree
 
 from ..utils.data_types import Dims, DataVar, DataArray, DataSet, Data, DataVarBound
 
@@ -23,12 +29,25 @@ class Transformer(BaseEstimator, TransformerMixin, ABC):
         self.feature_name = feature_name
 
     @abstractmethod
+    def get_serialization_attrs(self) -> Dict:
+        """Return a dictionary containing the attributes that need to be serialized
+        as part of a saved transformer.
+
+        There are limitations on the types of attributes that can be serialized.
+        Most simple types (e.g. int, float, str, bool, None) can be, as well as
+        DataArrays and dicts of DataArrays. Other nested types (e.g. lists of
+        DataArrays) will likely fail.
+
+        """
+        return dict()
+
+    @abstractmethod
     def fit(
         self,
         X: Data,
         sample_dims: Optional[Dims] = None,
         feature_dims: Optional[Dims] = None,
-        **kwargs
+        **kwargs,
     ) -> Self:
         """Fit transformer to data.
 
@@ -52,7 +71,7 @@ class Transformer(BaseEstimator, TransformerMixin, ABC):
         X: Data,
         sample_dims: Optional[Dims] = None,
         feature_dims: Optional[Dims] = None,
-        **kwargs
+        **kwargs,
     ) -> Data:
         return self.fit(X, sample_dims, feature_dims, **kwargs).transform(X)
 
@@ -71,3 +90,89 @@ class Transformer(BaseEstimator, TransformerMixin, ABC):
     @abstractmethod
     def inverse_transform_scores_unseen(self, X: DataArray) -> DataArray:
         return X
+
+    def _serialize_data(self, key: str, data: DataArray) -> DataSet:
+        # Make sure the DataArray has some name so we can create a string mapping
+        if data.name is None:
+            data.name = key
+
+        multiindexes = {}
+        if data.name in data.coords:
+            # Create coords-based datasets and note multiindexes
+            if isinstance(data.to_index(), pd.MultiIndex):
+                multiindexes[data.name] = [n for n in data.to_index().names]
+            ds = xr.Dataset(coords={data.name: data})
+        else:
+            # Create data-based datasets
+            ds = xr.Dataset(data_vars={data.name: data})
+
+        # Drop multiindexes and record for later
+        ds = ds.reset_index(list(multiindexes.keys()))
+        ds.attrs["multiindexes"] = multiindexes
+        ds.attrs["name_map"] = {key: data.name}
+
+        return ds
+
+    def serialize(self) -> DataTree:
+        """Serialize a transformer to a DataTree."""
+        from datatree import DataTree
+
+        dt = DataTree()
+        params = self.get_params()
+        attrs = self.get_serialization_attrs()
+
+        # Set initialization params as tree level attrs
+        dt.attrs["params"] = params
+
+        # Serialize each transformer attribute
+        for key, attr in attrs.items():
+            if isinstance(attr, xr.DataArray):
+                # attach data to data_vars or coords
+                ds = self._serialize_data(key, attr)
+                dt[key] = DataTree(name=key, data=ds)
+                dt.attrs[key] = "_is_node"
+            elif isinstance(attr, dict) and any(
+                [isinstance(val, xr.DataArray) for val in attr.values()]
+            ):
+                # attach dict of data as branching tree
+                dt_attr = DataTree()
[…]
