### 2023-11-07 7c1dc16
feat!: inverse_transform scores arg
