+++ tests/models/test_orthogonality.py
@@ -697,7 +697,8 @@ def test_eof_inverse_transform(dim, use_coslat, mock_data_array):
     data = mock_data_array
     model = EOF(n_modes=19, standardize=True, use_coslat=use_coslat)
     model.fit(data, dim=dim)
-    data_rec = model.inverse_transform(mode=slice(1, None))
+    scores = model.data["scores"]
+    data_rec = model.inverse_transform(scores)
     r2 = r2_score(data, data_rec, dim=dim)
     r2 = r2.mean()
     # Choose a threshold of 0.95; a bit arbitrary
@@ -718,7 +719,8 @@ def test_ceof_inverse_transform(dim, use_coslat, mock_data_array):
     data = mock_data_array
     model = ComplexEOF(n_modes=19, standardize=True, use_coslat=use_coslat)
     model.fit(data, dim=dim)
-    data_rec = model.inverse_transform(mode=slice(1, None)).real
+    scores = model.data["scores"]
+    data_rec = model.inverse_transform(scores).real
     r2 = r2_score(data, data_rec, dim=dim)
     r2 = r2.mean()
     # Choose a threshold of 0.95; a bit arbitrary
@@ -744,7 +746,8 @@ def test_reof_inverse_transform(dim, use_coslat, power, mock_data_array):
     model.fit(data, dim=dim)
     rot = EOFRotator(n_modes=19, power=power)
     rot.fit(model)
-    data_rec = rot.inverse_transform(mode=slice(1, None)).real
+    scores = rot.data["scores"]
+    data_rec = rot.inverse_transform(scores).real
     r2 = r2_score(data, data_rec, dim=dim)
     r2 = r2.mean()
     # Choose a threshold of 0.95; a bit arbitrary
@@ -772,7 +775,8 @@ def test_creof_inverse_transform(dim, use_coslat, power, mock_data_array):
     model.fit(data, dim=dim)
     rot = ComplexEOFRotator(n_modes=19, power=power)
     rot.fit(model)
-    data_rec = rot.inverse_transform(mode=slice(1, None)).real
+    scores = rot.data["scores"]
+    data_rec = rot.inverse_transform(scores).real
     r2 = r2_score(data, data_rec, dim=dim)
     r2 = r2.mean()
     # Choose a threshold of 0.95; a bit arbitrary
@@ -796,7 +800,9 @@ def test_mca_inverse_transform(dim, use_coslat, mock_data_array):
     data2 = data1.copy() ** 2
     model = MCA(n_modes=19, standardize=True, use_coslat=use_coslat)
     model.fit(data1, data2, dim=dim)
-    data1_rec, data2_rec = model.inverse_transform(mode=slice(1, None))
+    scores1 = model.data["scores1"]
+    scores2 = model.data["scores2"]
+    data1_rec, data2_rec = model.inverse_transform(scores1, scores2)
     r2_1 = r2_score(data1, data1_rec, dim=dim)
     r2_2 = r2_score(data2, data2_rec, dim=dim)
     r2_1 = r2_1.mean()
@@ -825,7 +831,9 @@ def test_cmca_inverse_transform(dim, use_coslat, mock_data_array):
     data2 = data1.copy() ** 2
     model = ComplexMCA(n_modes=19, standardize=True, use_coslat=use_coslat)
     model.fit(data1, data2, dim=dim)
-    data1_rec, data2_rec = model.inverse_transform(mode=slice(1, None))
+    scores1 = model.data["scores1"]
+    scores2 = model.data["scores2"]
+    data1_rec, data2_rec = model.inverse_transform(scores1, scores2)
[…]
