### 2023-11-07 #6009 @matthewfeickert
> We don't usually overlap but I'm hoping you'd be a good fit for editing this submission from the stats/math side of things. What do you think?

Sorry @kthyng I'm just seeing this now (I'm behind on email and GitHub notifications while applying for faculty positions). Hm, this is certainly outside of my wheelhouse subject wise, so if there is another topical editor that is a better fit on the subject matter I would go with them, but if not then I could take this on.
