+++ xeofs/models/_base_cross_model.py
@@ -255,7 +255,7 @@ class _BaseCrossModel(ABC):
         """Get the model parameters."""
         return self._params
 
-    def serialize(self, save_data: bool = False) -> DataSet:
+    def serialize(self, save_data: bool = False) -> DataTree:
         """Serialize a complete model with its preprocessors."""
         from datatree import DataTree
 
@@ -274,7 +274,7 @@ class _BaseCrossModel(ABC):
         # Store the DataContainer items as data_vars, and the model parameters as global attrs
         ds_model = xr.Dataset(data, attrs=dict(params=self.get_params()))
         # Set as the root node of the tree
-        dt = DataTree(name="model", data=ds_model)
+        dt = DataTree(data=ds_model, name=type(self).__name__)
 
         # Retrieve the tree representation of the preprocessor
         dt["preprocessor1"] = self.preprocessor1.serialize_all()
