### 2023-11-07 039f568
docs: add basic entry on serialization
