### 2023-11-07 7dabaa3
build: make zarr, xarray-datatree non-optional
