+++ .github/workflows/ci.yml
@@ -38,7 +38,7 @@ jobs:
         run: poetry --version
 
       - name: Install Project Dependencies
-        run: python -m poetry install --with dev,io
+        run: python -m poetry install --with dev
 
       - name: Lint Codebase with flake8
         run: |
