### 2023-11-07 21b566a
docs: add example for model save, load usage
