### 2023-11-07 e08b408
chore: cleanup from optional DataTree dependency
