### 2023-11-07 69df4d9
feat: serialization methods (#103)
Co-authored-by: Niclas Rieger <niclasrieger@gmail.com>
