+++ xeofs/preprocessing/scaler.py
@@ -1,4 +1,4 @@
-from typing import Optional
+from typing import Optional, Dict
 from typing_extensions import Self
 
 import numpy as np
@@ -39,6 +39,19 @@ class Scaler(Transformer):
         self.with_center = with_center
         self.with_std = with_std
         self.with_coslat = with_coslat
+
+        self.mean_ = xr.DataArray(name="mean_")
+        self.std_ = xr.DataArray(name="std_")
+        self.coslat_weights_ = xr.DataArray(name="coslat_weights_")
+        self.weights_ = xr.DataArray(name="weights_")
+
+    def get_serialization_attrs(self) -> Dict:
+        return dict(
+            mean_=self.mean_,
+            std_=self.std_,
+            coslat_weights_=self.coslat_weights_,
+            weights_=self.weights_,
+        )
 
     def _verify_input(self, X, name: str):
         if not isinstance(X, (xr.DataArray, xr.Dataset)):
@@ -173,163 +186,3 @@ class Scaler(Transformer):
 
     def inverse_transform_scores_unseen(self, X: DataArray) -> DataArray:
         return X
-
-
-# class DataListScaler(Scaler):
-#     """Scale a list of xr.DataArray along sample dimensions.
-
-#     Scaling includes (i) removing the mean and, optionally, (ii) dividing by the standard deviation,
-#     (iii) multiplying by the square root of cosine of latitude weights (area weighting; coslat weighting),
-#     and (iv) multiplying by additional user-defined weights.
-
-#     Parameters
-#     ----------
-#     with_std : bool, default=True
-#         If True, the data is divided by the standard deviation.
-#     with_coslat : bool, default=False
-#         If True, the data is multiplied by the square root of cosine of latitude weights.
-#     with_weights : bool, default=False
-#         If True, the data is multiplied by additional user-defined weights.
-
-#     """
-
-#     def __init__(self, with_std=False, with_coslat=False):
-#         super().__init__(with_std=with_std, with_coslat=with_coslat)
-#         self.scalers = []
-
-#     def _verify_input(self, data, name: str):
-#         """Verify that the input data is a list of DataArrays.
-
-#         Parameters
-#         ----------
-#         data : list of xarray.DataArray
-#             Data to be checked.
-
-#         """
-#         if not isinstance(data, list):
-#             raise TypeError(f"{name} must be a list of xarray DataArrays or Datasets")
-#         if not all(isinstance(da, (xr.DataArray, xr.Dataset)) for da in data):
-#             raise TypeError(f"{name} must be a list of xarray DataArrays or Datasets")
-
-#     def fit(
-#         self,
-#         data: List[Data],
-#         sample_dims: Dims,
-#         feature_dims_list: DimsList,
-#         weights: Optional[List[Data] | Data] = None,
-#     ) -> Self:
-#         """Fit the scaler to the data.
-
-#         Parameters
-#         ----------
-#         data : list of xarray.DataArray
-#             Data to be scaled.
-#         sample_dims : hashable or sequence of hashable
-#             Dimensions along which the data is considered to be a sample.
-#         feature_dims_list : list of hashable or list of sequence of hashable
-#             List of dimensions along which the data is considered to be a feature.
-#         weights : list of xarray.DataArray, optional
-#             List of weights to be applied to the data. Must have the same dimensions as the data.
-
-#         """
-#         self._verify_input(data, "data")
-
-#         # Check input
-#         if not isinstance(feature_dims_list, list):
-#             err_message = "feature dims must be a list of the feature dimensions of each DataArray, "
-#             err_message += 'e.g. [("lon", "lat"), ("lon")]'
-#             raise TypeError(err_message)
-
-#         # Sample dimensions are the same for all data arrays
[…]
