+++ xeofs/preprocessing/stacker.py
@@ -1,18 +1,18 @@
-from abc import abstractmethod
-from typing import List, Optional, Type
+from typing import Dict
 from typing_extensions import Self
 
-import numpy as np
 import pandas as pd
 import xarray as xr
 
 from .transformer import Transformer
 from ..utils.data_types import Dims, DataArray, DataSet, Data, DataVar, DataVarBound
-from ..utils.sanity_checks import convert_to_dim_type
 
 
 class Stacker(Transformer):
-    """Converts a DataArray of any dimensionality into a 2D structure.
+    """Converts an xarray DataArray or Dataset of any dimensionality into a 2D DataArray.
+
+    The new DataArray will have two dimensions: `sample` and `feature`.
+    The dimensions of the original data will be stacked along these two dimensions.
 
     Attributes
     ----------
@@ -21,9 +21,9 @@ class Stacker(Transformer):
     feature_dims : Sequence[Hashable]
         The dimensions of the data that will be stacked along the `feature` dimension.
     sample_name : str
-        The name of the sample dimension.
+        The name of the sample dimension (dim=0).
     feature_name : str
-        The name of the feature dimension.
+        The name of the feature dimension (dim=1).
     dims_in : Tuple[str]
         The dimensions of the input data.
     dims_out : Tuple[str]
@@ -47,11 +47,82 @@ class Stacker(Transformer):
         self.dims_out = tuple((sample_name, feature_name))
         self.dims_mapping = {}
         self.dims_mapping.update({d: tuple() for d in self.dims_out})
-
         self.coords_in = {}
         self.coords_out = {}
 
-    def _validate_matching_dimensions(self, X: Data):
+    def get_serialization_attrs(self) -> Dict:
+        return dict(
+            dims_in=self.dims_in,
+            dims_out=self.dims_out,
+            dims_mapping=self.dims_mapping,
+            coords_in=self.coords_in,
+            coords_out=self.coords_out,
+        )
+
+    def _validate_data_type(self, X: Data):
+        """Check that the data type is either DataArray or Dataset."""
+        if not isinstance(X, (xr.DataArray, xr.Dataset)):
+            raise TypeError(f"Invalid data type {type(X)}.")
+
+    def _validate_dimension_names(self, X, sample_dims, feature_dims):
+        """Check that the names of the sample and feature dimensions are not already present in the data."""
+        sample_name_in_data = self.sample_name in X.dims
+        feature_name_in_data = self.feature_name in X.dims
+
+        has_invalid_sample_name = (
+            True if (len(sample_dims) > 1) and sample_name_in_data else False
+        )
+
+        match X:
+            case xr.DataArray():
+                has_invalid_feature_name = (
+                    True if (len(feature_dims) > 1) and feature_name_in_data else False
+                )
+            case xr.Dataset():
+                has_invalid_feature_name = True if feature_name_in_data else False
+            case _:
+                raise TypeError(f"Invalid data type {type(X)}.")
+
+        if has_invalid_sample_name:
+            err_msg = f"Name of sample dimension ({self.sample_name}) is already present in data. Please use another name."
+            raise ValueError(err_msg)
+
+        if has_invalid_feature_name:
+            err_msg = f"Name of feature dimension ({self.feature_name}) is already present in data. Please use another name."
+            raise ValueError(err_msg)
+
+    def _validate_dims(self, X: Data, sample_dims, feature_dims):
+        invalid_sample_dims = True if len(sample_dims) < 1 else False
+        invalid_feature_dims = True if len(feature_dims) < 1 else False
+
+        if invalid_sample_dims:
+            raise ValueError(f"Sample dimension must not be empty.")
+        if invalid_feature_dims:
+            match X:
+                case xr.DataArray():
[…]
