### 2023-11-07 ecb92b1
2.0.0
Automatically generated by python-semantic-release
