### 2023-11-07 #6009 @kthyng
Hi @samhforbes! I am looking for an editor in the signal processing/time series analysis sort of area. Are you up for editing this submission?
