### 2023-11-07 8cf91ea
2.0.1
Automatically generated by python-semantic-release
