### 2023-11-08 0f95d04
fix: safe default zarr write mode
