+++ xeofs/models/_base_model.py
@@ -357,6 +357,7 @@ class _BaseModel(ABC):
     def save(
         self,
         path: str,
+        overwrite: bool = False,
         save_data: bool = False,
         **kwargs,
     ):
@@ -366,6 +367,8 @@ class _BaseModel(ABC):
         ----------
         path : str
             Path to save the model zarr store.
+        overwrite: bool, default=False
+            Whether or not to overwrite the existing path if it already exists.
         save_data : str
             Whether or not to save the full input data along with the fitted components.
         **kwargs
@@ -379,7 +382,9 @@ class _BaseModel(ABC):
         if hasattr(self, "model"):
             dt["model"] = self.model.serialize(save_data=save_data)
 
-        dt.to_zarr(path, **kwargs)
+        write_mode = "w" if overwrite else "w-"
+
+        dt.to_zarr(path, mode=write_mode, **kwargs)
 
     @classmethod
     def deserialize(cls, dt: DataTree) -> Self:
