### 2023-11-08 1f38a5b
2.0.2
Automatically generated by python-semantic-release
