+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "xeofs"
-version = "2.0.1"
+version = "2.0.2"
 description = "Comprehensive EOF analysis in Python with xarray: A versatile, multidimensional, and scalable tool for advanced climate data analysis"
 authors = ["Niclas Rieger <niclasrieger@gmail.com>"]
 license = "MIT"
