### 2023-11-08 b9b56f0
feat: inverse_transform for EEOF
