+++ docs/auto_examples/1single/plot_eeof trend.rst
@@ -1,0 +1,276 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/1single/plot_eeof trend.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_1single_plot_eeof trend.py>`
+        to download the full example code
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_1single_plot_eeof trend.py:
+
+
+Removing nonlinear trends with EEOF analysis
+============================================
+
+This example demonstrates how to use Extended EOF (EEOF) analysis 
+in order to remove nonlinear trends from a dataset.
+
+Let's begin by setting up the required packages and fetching the data:
+
+.. GENERATED FROM PYTHON SOURCE LINES 10-17
+
+.. code-block:: default
+
+
+    import xarray as xr
+    import xeofs as xe
+    import matplotlib.pyplot as plt
+
+    xr.set_options(display_expand_data=False)
+
+
+
+
+
+.. rst-class:: sphx-glr-script-out
+
+ .. code-block:: none
+
+
+    <xarray.core.options.set_options object at 0x7ff3615fb350>
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 18-19
+
+Load the tutorial data.
+
+.. GENERATED FROM PYTHON SOURCE LINES 19-24
+
+.. code-block:: default
+
+
+    sst = xr.tutorial.open_dataset("ersstv5").sst
+    sst = sst.groupby("time.month") - sst.groupby("time.month").mean("time")
+
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 25-38
+
+Prior to conducting the EEOF analysis, it's essential to determine the
+structure of the lagged covariance matrix. This entails defining the time
+delay ``tau`` and the ``embedding`` dimension. The former signifies the
+interval between the original and lagged time series, while the latter
+dictates the number of time-lagged copies in the delay-coordinate space,
+representing the system's dynamics.
+For illustration, using ``tau=4`` and ``embedding=40``, we generate 40
+delayed versions of the time series, each offset by 4 time steps, resulting
+in a maximum shift of ``tau x embedding = 160``. Given our dataset's
+6-hour intervals, tau = 4 translates to a 24-hour shift.
+It's obvious that this way of constructing the lagged covariance matrix
+and subsequently decomposing it can be computationally expensive. For example,
+given our dataset's dimensions,
+
+.. GENERATED FROM PYTHON SOURCE LINES 38-44
+
+.. code-block:: default
+
+
+    eof = xe.models.EOF(n_modes=10)
+    eof.fit(sst, dim="time")
+    scores = eof.scores()
+    components = eof.components()
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 45-51
+
+.. code-block:: default
+
+
+    fig, ax = plt.subplots(1, 2, figsize=(10, 5))
+    scores.sel(mode=1).plot(ax=ax[0])
+    components.sel(mode=1).plot(ax=ax[1])
[…]
