+++ docs/auto_examples/index.rst
@@ -96,6 +96,23 @@ Here you can find some examples of how to use the library.
 
 .. raw:: html
 
+    <div class="sphx-glr-thumbcontainer" tooltip="This example demonstrates how to use Extended EOF (EEOF) analysis  in order to remove nonlinear...">
+
+.. only:: html
+
+  .. image:: /auto_examples/1single/images/thumb/sphx_glr_plot_eeof trend_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_auto_examples_1single_plot_eeof trend.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Removing nonlinear trends with EEOF analysis</div>
+    </div>
+
+
+.. raw:: html
+
     <div class="sphx-glr-thumbcontainer" tooltip="Multivariate EOF analysis.">
 
 .. only:: html
