+++ docs/_autosummary/xeofs.models.ComplexMCARotator.rst
@@ -22,14 +22,18 @@
       ~ComplexMCARotator.components_phase
       ~ComplexMCARotator.compute
       ~ComplexMCARotator.covariance_fraction
+      ~ComplexMCARotator.deserialize
       ~ComplexMCARotator.fit
       ~ComplexMCARotator.get_params
       ~ComplexMCARotator.heterogeneous_patterns
       ~ComplexMCARotator.homogeneous_patterns
       ~ComplexMCARotator.inverse_transform
+      ~ComplexMCARotator.load
+      ~ComplexMCARotator.save
       ~ComplexMCARotator.scores
       ~ComplexMCARotator.scores_amplitude
       ~ComplexMCARotator.scores_phase
+      ~ComplexMCARotator.serialize
       ~ComplexMCARotator.singular_values
       ~ComplexMCARotator.squared_covariance
       ~ComplexMCARotator.squared_covariance_fraction

[…]
