### 2023-11-09 256a415
perf: eliminate unnecessary eagar computations
