+++ xeofs/preprocessing/scaler.py
@@ -99,18 +99,18 @@ class Scaler(Transformer):
 
         # Scaling parameters are computed along sample dimensions
         if params["with_center"]:
-            self.mean_: DataVar = X.mean(self.sample_dims).compute()
+            self.mean_: DataVar = X.mean(self.sample_dims)
 
         if params["with_std"]:
-            self.std_: DataVar = X.std(self.sample_dims).compute()
+            self.std_: DataVar = X.std(self.sample_dims)
 
         if params["with_coslat"]:
             self.coslat_weights_: DataVar = compute_sqrt_cos_lat_weights(
                 data=X, feature_dims=self.feature_dims
-            ).compute()
+            )
 
         # Convert None weights to ones
-        self.weights_: DataVar = self._process_weights(X, weights).compute()
+        self.weights_: DataVar = self._process_weights(X, weights)
 
         return self
 
