### 2023-11-09 661d006
perf: lazy import njit functions
