+++ xeofs/models/gwpca.py
@@ -18,11 +18,8 @@ from .eof import EOF
 import numpy as np
 import xarray as xr
 
-from tqdm import trange
-import numba
-from numba import prange
-from ..utils.distance_metrics import distance_nb, VALID_METRICS
-from ..utils.kernels import kernel_weights_nb, VALID_KERNELS
+from ..utils.distance_metrics import VALID_METRICS
+from ..utils.kernels import VALID_KERNELS
 
 
 class GWPCA(_BaseModel):
@@ -134,6 +131,9 @@ class GWPCA(_BaseModel):
         self.kernel = kernel
 
     def _fit_algorithm(self, X: DataArray) -> Self:
+        # Hide numba imports here to greatly speed up module import time
+        from ..utils.numba_utils import _local_pcas
+
         # Convert Dask arrays
         if not isinstance(X.data, np.ndarray):
             print(
@@ -253,184 +253,3 @@ class GWPCA(_BaseModel):
 
     def _inverse_transform_algorithm(self, scores: DataArray) -> DataArray:
         raise NotImplementedError("GWPCA does not support inverse_transform() yet.")
-
-
-# Additional utility functions for local PCA
-# =============================================================================
-
-
-@numba.njit(fastmath=True, parallel=True)
-def _local_pcas(X, xy, n_modes, metric, kernel, bandwidth):
-    """Perform local PCA on each sample.
-
-    Parameters
-    ----------
-    X: ndarray
-        Input data with shape (n_samples, n_features)
-    xy: ndarray
-        Sample coordinates with shape (n_samples, 2)
-    n_modes: int
-        Number of modes to calculate.
-    metric: str
-        Distance metric to use. Great circle distance (`haversine`) is always expressed in kilometers.
-        All other distance metrics are reported in the unit of the input data.
-        See scipy.spatial.distance.cdist for a list of available metrics.
-    kernel: str
-        Kernel function to use. Must be one of ['bisquare', 'gaussian', 'exponential'].
-    bandwidth: float
-        Bandwidth of the kernel function.
-
-    Returns
-    -------
-    ndarray
-        Array of local components with shape (n_samples, n_features, n_modes)
-    ndarray
-        Array of local explained variance with shape (n_samples, n_modes)
-    ndarray
-        Array of total variance with shape (n_samples,)
-
-    """
-    n_samples = X.shape[0]
-    n_features = X.shape[1]
-    Vs = np.empty((n_samples, n_features, n_modes))
-    exp_var = np.empty((n_samples, n_modes))
-    tot_var = np.empty(n_samples)
-    for i in prange(n_samples):
-        dist = distance_nb(xy, xy[i], metric=metric)
-        weights = kernel_weights_nb(dist, bandwidth, kernel)
-        valid_data = weights > 0
-
-        weights = weights[valid_data]
-        x = X[valid_data]
-
-        wmean = _wmean_axis0(x, weights)
-        x -= wmean
-
-        sqrt_weights = np.sqrt(weights)
-        x = _weigh_columns(x, sqrt_weights)
-
-        Ui, si, ViT = np.linalg.svd(x, full_matrices=False)
-        # Renormalize singular values
-        si = si**2 / weights.sum()
-        ti = si.sum()
-
-        si = si[:n_modes]
-        ViT = ViT[:n_modes]
-        Vi = ViT.T
-
-        Vs[i] = Vi
-        exp_var[i, : len(si)] = si
-        tot_var[i] = ti
-
-    return Vs, exp_var, tot_var
-
-
-@numba.njit(fastmath=True)
-def _wmean_axis0(X, weights):
-    """Compute weighted mean along axis 0.
-
[…]
