### 2023-11-09 6a25c5c
docs: add EEOF trend removal example
