+++ docs/auto_examples/1single/index.rst
@@ -82,14 +82,14 @@
 
 .. raw:: html
 
-    <div class="sphx-glr-thumbcontainer" tooltip="This example demonstrates how to use Extended EOF (EEOF) analysis  in order to remove nonlinear...">
+    <div class="sphx-glr-thumbcontainer" tooltip="This tutorial illustrates the application of Extended EOF (EEOF) analysis to isolate and remove...">
 
 .. only:: html
 
-  .. image:: /auto_examples/1single/images/thumb/sphx_glr_plot_eeof trend_thumb.png
+  .. image:: /auto_examples/1single/images/thumb/sphx_glr_plot_eeof_trend_thumb.png
     :alt:
 
-  :ref:`sphx_glr_auto_examples_1single_plot_eeof trend.py`
+  :ref:`sphx_glr_auto_examples_1single_plot_eeof_trend.py`
 
 .. raw:: html
 
@@ -194,7 +194,7 @@
    /auto_examples/1single/plot_eof-tmode
    /auto_examples/1single/plot_complex_eof
    /auto_examples/1single/plot_eof-smode
-   /auto_examples/1single/plot_eeof trend
+   /auto_examples/1single/plot_eeof_trend
    /auto_examples/1single/plot_multivariate-eof
    /auto_examples/1single/plot_mreof
    /auto_examples/1single/plot_rotated_eof
