+++ docs/auto_examples/index.rst
@@ -96,14 +96,14 @@ Here you can find some examples of how to use the library.
 
 .. raw:: html
 
-    <div class="sphx-glr-thumbcontainer" tooltip="This example demonstrates how to use Extended EOF (EEOF) analysis  in order to remove nonlinear...">
+    <div class="sphx-glr-thumbcontainer" tooltip="This tutorial illustrates the application of Extended EOF (EEOF) analysis to isolate and remove...">
 
 .. only:: html
 
-  .. image:: /auto_examples/1single/images/thumb/sphx_glr_plot_eeof trend_thumb.png
+  .. image:: /auto_examples/1single/images/thumb/sphx_glr_plot_eeof_trend_thumb.png
     :alt:
 
-  :ref:`sphx_glr_auto_examples_1single_plot_eeof trend.py`
+  :ref:`sphx_glr_auto_examples_1single_plot_eeof_trend.py`
 
 .. raw:: html
 
