### 2023-11-09 ab6e3cc
feat: add EEOF inverse_transform() (#111)
