+++ docs/auto_examples/1single/plot_eeof_trend.rst
@@ -1,0 +1,289 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/1single/plot_eeof_trend.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_1single_plot_eeof_trend.py>`
+        to download the full example code
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_1single_plot_eeof_trend.py:
+
+
+Removing nonlinear trends with EEOF analysis
+============================================
+
+This tutorial illustrates the application of Extended EOF (EEOF) analysis
+to isolate and remove nonlinear trends within a dataset.
+
+Let's begin by setting up the required packages and fetching the data.
+
+.. GENERATED FROM PYTHON SOURCE LINES 10-17
+
+.. code-block:: default
+
+
+    import xarray as xr
+    import xeofs as xe
+    import matplotlib.pyplot as plt
+
+    xr.set_options(display_expand_data=False)
+
+
+
+
+
+.. rst-class:: sphx-glr-script-out
+
+ .. code-block:: none
+
+
+    <xarray.core.options.set_options object at 0x7f4de4c12bd0>
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 18-21
+
+We load the sea surface temperature (SST) data from the xarray tutorial.
+The dataset consists of monthly averages from 1970 to 2021. To ensure the seasonal
+cycle doesn't overshadow the analysis, we remove the monthly climatologies.
+
+.. GENERATED FROM PYTHON SOURCE LINES 21-26
+
+.. code-block:: default
+
+
+    sst = xr.tutorial.open_dataset("ersstv5").sst
+    sst = sst.groupby("time.month") - sst.groupby("time.month").mean("time")
+
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 27-28
+
+We start by performing a standard EOF analysis on the dataset.
+
+.. GENERATED FROM PYTHON SOURCE LINES 28-34
+
+.. code-block:: default
+
+
+    eof = xe.models.EOF(n_modes=10)
+    eof.fit(sst, dim="time")
+    scores = eof.scores()
+    components = eof.components()
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 35-39
+
+We immediately see that the first mode represents the global warming trend.
+Yet, the signal is somewhat muddled by short-term and year-to-year variations.
+Note the pronounced spikes around 1998 and 2016, hinting at the leakage of
+ENSO signatures into this mode.
+
+.. GENERATED FROM PYTHON SOURCE LINES 39-45
+
+.. code-block:: default
+
+
+    fig, ax = plt.subplots(1, 2, figsize=(10, 5))
+    scores.sel(mode=1).plot(ax=ax[0])
+    components.sel(mode=1).plot(ax=ax[1])
+
+
+
+
+
+.. image-sg:: /auto_examples/1single/images/sphx_glr_plot_eeof_trend_001.png
+   :alt: mode = 1, mode = 1
+   :srcset: /auto_examples/1single/images/sphx_glr_plot_eeof_trend_001.png
+   :class: sphx-glr-single-img
+
+
[…]
