+++ docs/auto_examples/1single/sg_execution_times.rst
@@ -6,14 +6,10 @@
 
 Computation times
 =================
-**00:30.679** total execution time for **auto_examples_1single** files:
+**00:28.072** total execution time for **auto_examples_1single** files:
 
 +-----------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_examples_1single_plot_rotated_eof.py` (``plot_rotated_eof.py``)           | 00:18.587 | 0.0 MB |
-+-----------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_examples_1single_plot_mreof.py` (``plot_mreof.py``)                       | 00:06.486 | 0.0 MB |
-+-----------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_examples_1single_plot_weighted-eof.py` (``plot_weighted-eof.py``)         | 00:05.606 | 0.0 MB |
+| :ref:`sphx_glr_auto_examples_1single_plot_eeof_trend.py` (``plot_eeof_trend.py``)             | 00:28.072 | 0.0 MB |
 +-----------------------------------------------------------------------------------------------+-----------+--------+
 | :ref:`sphx_glr_auto_examples_1single_plot_complex_eof.py` (``plot_complex_eof.py``)           | 00:00.000 | 0.0 MB |
 +-----------------------------------------------------------------------------------------------+-----------+--------+
@@ -25,5 +21,11 @@ Computation times
 +-----------------------------------------------------------------------------------------------+-----------+--------+
 | :ref:`sphx_glr_auto_examples_1single_plot_gwpca.py` (``plot_gwpca.py``)                       | 00:00.000 | 0.0 MB |
 +-----------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_auto_examples_1single_plot_mreof.py` (``plot_mreof.py``)                       | 00:00.000 | 0.0 MB |
++-----------------------------------------------------------------------------------------------+-----------+--------+
 | :ref:`sphx_glr_auto_examples_1single_plot_multivariate-eof.py` (``plot_multivariate-eof.py``) | 00:00.000 | 0.0 MB |
++-----------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_auto_examples_1single_plot_rotated_eof.py` (``plot_rotated_eof.py``)           | 00:00.000 | 0.0 MB |
++-----------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_auto_examples_1single_plot_weighted-eof.py` (``plot_weighted-eof.py``)         | 00:00.000 | 0.0 MB |
 +-----------------------------------------------------------------------------------------------+-----------+--------+
