+++ docs/_autosummary/xeofs.models.EOFRotator.rst
@@ -19,13 +19,17 @@
       ~EOFRotator.__init__
       ~EOFRotator.components
       ~EOFRotator.compute
+      ~EOFRotator.deserialize
       ~EOFRotator.explained_variance
       ~EOFRotator.explained_variance_ratio
       ~EOFRotator.fit
       ~EOFRotator.fit_transform
       ~EOFRotator.get_params
       ~EOFRotator.inverse_transform
+      ~EOFRotator.load
+      ~EOFRotator.save
       ~EOFRotator.scores
+      ~EOFRotator.serialize
       ~EOFRotator.singular_values
       ~EOFRotator.transform
    
