### 2023-11-09 4e3c2e5
perf: delay additional computes in rotators
