+++ xeofs/utils/xarray_utils.py
@@ -269,3 +269,64 @@ def _np_sqrt_cos_lat_weights(data):
 
     """
     return np.sqrt(np.cos(np.deg2rad(data)).clip(0, 1))
+
+
+def get_deterministic_sign_multiplier(data: DataArray, dim: str) -> DataArray:
+    """Compute a sign multiplier that ensures deterministic output.
+
+    Uses a method standard to other SVD implementations where we ensure that
+    the maximum absolute value in the passed data matrix is positive
+    when multipled by the sign multiplier. This creates deterministic output.
+
+    Parameters:
+    ------------
+    data: DataArray
+        Input data to determine sorting order.
+    dim: str
+        Dimension along which to compute the sign multiplier.
+
+    Returns:
+    ---------
+    sign_multiplier: DataArray
+        Sign multiplier that ensures deterministic output.
+    """
+    # This method is carefully constructed to avoid idexing ops
+    # so we can execute this lazily on dask arrays
+    min_max = xr.concat([data.max(dim), data.min(dim)], dim="sign")
+    min_max = min_max.assign_coords(sign=[1, -1])
+    sign_multiplier = np.abs(min_max).idxmax("sign")
+    # Drop all dimensions except 'mode' so that the index is clean
+    for dim, coords in sign_multiplier.coords.items():
+        if dim != "mode":
+            sign_multiplier = sign_multiplier.drop(dim)
+    return sign_multiplier
+
+
+def argsort_dask(data: Data, dim: str) -> Data:
+    """Apply argsort to a dask-backed array.
+
+    This is a workaround because dask does not yet implement a chunk-aware
+    version of argsort. Therefore we have to force rechunking to a single
+    chunk along the sorting dimension and then apply numpy argsort. This
+    should be used with an understanding that it may produce large memory
+    usage if the passed array is chunked finely along the sorting dimension.
+
+    Parameters:
+    ------------
+    data: Data
+        Input data to sort.
+    dim: str
+        Dimension along which to sort.
+
+    Returns:
+    ---------
+    sorted_idx: Data
+        Indices that would sort the data.
+    """
+    return xr.apply_ufunc(
+        np.argsort,
+        data.chunk({dim: -1}),
+        input_core_dims=[[dim]],
+        output_core_dims=[[dim]],
+        dask="parallelized",
+    )

[…]
