### 2023-11-10 31f4da1
perf: delay rotator sorting as post-compute method
