### 2023-11-11 c1c5d82
perf: optional lazy exection through rotators
