### 2023-11-12 1400cab
perf: simultaneous compute of all dask collections
