+++ xeofs/models/_base_cross_model.py
@@ -3,14 +3,16 @@ from typing_extensions import Self
 from abc import ABC, abstractmethod
 from datetime import datetime
 
+import dask
 import xarray as xr
 from datatree import DataTree, open_datatree
+from dask.diagnostics.progress import ProgressBar
 
 from .eof import EOF
 from ..preprocessing.preprocessor import Preprocessor
 from ..data_container import DataContainer
 from ..utils.data_types import DataObject, DataArray, DataSet
-from ..utils.xarray_utils import convert_to_dim_type
+from ..utils.xarray_utils import convert_to_dim_type, data_is_dask
 from ..utils.sanity_checks import validate_input_type
 from .._version import __version__
 
@@ -125,6 +127,13 @@ class _BaseCrossModel(ABC):
             EOF(n_modes=n_pca_modes, compute=self._params["compute"])
             if n_pca_modes
             else None
+        )
+
+    def get_serialization_attrs(self) -> Dict:
+        return dict(
+            data=self.data,
+            preprocessor1=self.preprocessor1,
+            preprocessor2=self.preprocessor2,
         )
 
     def fit(
@@ -262,7 +271,43 @@ class _BaseCrossModel(ABC):
             Whether or not to provide additional information about the computing progress.
 
         """
-        self.data.compute(verbose=verbose)
+        dt = self.serialize()
+        data_objs = {
+            k: v
+            for k, v in dt.to_dict().items()
+            if data_is_dask(v) and v.attrs.get("allow_compute", True)
+        }
+
+        # Compute all dask collections simultaneously
+        if verbose:
+            with ProgressBar():
+                (data_objs,) = dask.compute(data_objs)
+        else:
+            (data_objs,) = dask.compute(data_objs)
+
+        # Reassign all computed arrays
+        for key, data in data_objs.items():
+            path_elems = key.strip("/").split("/")
+            parent = self
+            for elem in path_elems[:-1]:
+                if elem.isdigit():
+                    parent = parent[int(elem)]
+                else:
+                    parent = getattr(parent, elem)
+
+            if isinstance(data, xr.Dataset):
+                mapping = data.attrs.get("name_map", {}).get(
+                    path_elems[-1], path_elems[-1]
+                )
+                da = data[mapping]
+            else:
+                da = data
+
+            if isinstance(parent, dict):
+                parent[path_elems[-1]] = da
+            else:
+                setattr(parent, path_elems[-1], da)
+
         self._post_compute()
 
     def _post_compute(self):
@@ -274,31 +319,14 @@ class _BaseCrossModel(ABC):
 
     def serialize(self, save_data: bool = False) -> DataTree:
         """Serialize a complete model with its preprocessors."""
-        data = {}
-        for key, x in self.data.items():
-            if self.data._allow_compute[key] or save_data:
-                data[key] = x.assign_attrs(
-                    {"allow_compute": self.data._allow_compute[key]}
-                )
-            else:
-                # create an empty placeholder array
-                data[key] = xr.DataArray().assign_attrs(
-                    {"allow_compute": False, "placeholder": True}
-                )
+        # Create a root node for this object with its params as attrs
+        ds_root = xr.Dataset(attrs=dict(params=self.get_params()))
+        dt = DataTree(data=ds_root, name=type(self).__name__)
 
-        # Store the DataContainer items as data_vars, and the model parameters as global attrs
-        ds_model = xr.Dataset(data, attrs=dict(params=self.get_params()))
-        # Set as the root node of the tree
-        dt = DataTree(data=ds_model, name=type(self).__name__)
-
-        # Retrieve the tree representation of the preprocessor
[…]
