### 2023-11-12 ef4509b
fix: clip std so fillna works
