+++ xeofs/preprocessing/scaler.py
@@ -102,7 +102,9 @@ class Scaler(Transformer):
             self.mean_: DataVar = X.mean(self.sample_dims).rename("mean_")
 
         if params["with_std"]:
-            self.std_: DataVar = X.std(self.sample_dims).rename("std_")
+            self.std_: DataVar = (
+                X.std(self.sample_dims).clip(min=np.finfo(X.dtype).eps).rename("std_")
+            )
 
         if params["with_coslat"]:
             self.coslat_weights_: DataVar = compute_sqrt_cos_lat_weights(
