### 2023-11-12 3b7f4b5
doc: usage example for lazy evaluation
