+++ docs/overview_2_features.rst
@@ -45,34 +45,55 @@ dimensions are *feature* dimensions to be expressed as latent variables.
 Dask Support
 ----------------
 
-If you handle large datasets that exceed memory capacity, you can pass a ``Dask`` ``xarray`` object to ``xeofs``. 
-Contrary to expectations, ``xeofs`` computes the matrix decomposition results, assuming the output can fit 
-into memory. This behavior is the default because deferring computations usually causes substantial 
-increase in computational time since any derivative of the decomposition will have to recompute the entire SVD decomposition. 
-
-.. code-block:: python
-
-  model = xe.models.EOF()
-  model.fit(data, dim="time")  # <- data is a Dask object
-  
-  model.components() # Returns a non-Dask object
-
-However, if you wish to postpone computation, you can achieve this by setting ``compute=False`` during model initialization:
-
-.. code-block:: python
-
-  model = xe.models.EOF(compute=False)
-  model.fit(data, dim="time")
-  
-  model.components() # Returns a Dask object
-
-  model.compute() # Computes all deferred objects
-
-  model.components() # Now, it's no longer a Dask object
+If you handle large datasets that exceed memory capacity, ``xeofs`` is designed to work with ``dask``-backed
+``xarray`` objects from end-to-end. By default, ``xeofs`` computes models eagerly by default, which in some
+cases can lead to better performance. However, it is also possible to build and fit models "lazily", meaning
+no computation will be carried out until the user calls ``.compute()``. To enable lazy computation, specify
+``compute=False`` when initializing the model.
 
 .. note::
 
-    Importantly, ``xeofs`` never computes the input data directly.
+    Importantly, ``xeofs`` never loads the input dataset(s) into memory.
+
+Lazy Evaluation
+---------------
+
+There are a few tricks, and features that need to be explicitly disabled for lazy evaluation to work. First
+is the ``check_nans`` option, which skips checking for full or isolated ``NaNs`` in the data. In this case,
+the user is responsible for ensuring that the data is free of ``NaNs`` by first applying e.g. ``.dropna()``
+or ``.fillna()``. Second is that lazy mode is incompatible with assessing the fit of a rotator class during
+evaluation, becaue the entire ``dask`` task graph must be built up front. Therefore, a lazy rotator model will
+run out to the full ``max_iter`` regardless of the specified ``rtol``. For that reason it is recommended to
+reduce the number of iterations.
+
+As an example, the following lazily creates a rotated EOF model for an 80GB dataset in about a second, which can
+then be evaluated later using ``.compute()``.
+
+.. code-block:: python
+
+  import dask.array as da
+  import numpy as np
+  import xarray as xr
+
+  from xeofs.models import EOF, EOFRotator
+
+  data = xr.DataArray(
+      da.random.random((10000, 1440, 720), chunks=(100, 100, 100)),
+      dims=["time", "lon", "lat"],
+      coords={
+          "time": xr.date_range("2000-01-01", periods=10000, freq="D"),
+          "lon": np.linspace(-180, 180, 1440),
+          "lat": np.linspace(-90, 90, 720),
+      },
+  )
+  model = EOF(n_modes=5, check_nans=False, compute=False)
+  model.fit(data, dim="time")
+
+  rotator = EOFRotator(compute=False, max_iter=20)
+  rotator.fit(model)
+
+  # Later, you can compute the model
+  # rotator.compute()
 
 Available Methods
 -----------------
@@ -102,7 +123,6 @@ fitted models to a portable format.
 
   # Later, you can load the model
   loaded_model = EOF.load("my_model.zarr")
-
 
 Input Data Compatibility
 ------------------------
[…]
