### 2023-11-12 3be02c4
fix: pass compute through to scaler
