+++ xeofs/preprocessing/scaler.py
@@ -34,11 +34,13 @@ class Scaler(Transformer):
         with_center: bool = True,
         with_std: bool = False,
         with_coslat: bool = False,
+        compute: bool = True,
     ):
         super().__init__()
         self.with_center = with_center
         self.with_std = with_std
         self.with_coslat = with_coslat
+        self.compute = compute
 
         self.mean_ = xr.DataArray(name="mean_")
         self.std_ = xr.DataArray(name="std_")
@@ -99,20 +101,26 @@ class Scaler(Transformer):
 
         # Scaling parameters are computed along sample dimensions
         if params["with_center"]:
-            self.mean_: DataVar = X.mean(self.sample_dims).rename("mean_")
+            self.mean_: DataVar = X.mean(self.sample_dims)
 
         if params["with_std"]:
-            self.std_: DataVar = (
-                X.std(self.sample_dims).clip(min=np.finfo(X.dtype).eps).rename("std_")
+            self.std_: DataVar = X.std(self.sample_dims).clip(
+                min=np.finfo(np.float32).eps
             )
 
         if params["with_coslat"]:
             self.coslat_weights_: DataVar = compute_sqrt_cos_lat_weights(
                 data=X, feature_dims=self.feature_dims
-            ).rename("coslat_weights_")
+            )
 
         # Convert None weights to ones
-        self.weights_: DataVar = self._process_weights(X, weights).rename("weights_")
+        self.weights_: DataVar = self._process_weights(X, weights)
+
+        if self.get_params()["compute"]:
+            self.mean_ = self.mean_.compute()
+            self.std_ = self.std_.compute()
+            self.coslat_weights_ = self.coslat_weights_.compute()
+            self.weights_ = self.weights_.compute()
 
         return self
 
