+++ pyproject.toml
@@ -23,12 +23,6 @@ statsmodels = ">=0.14.0"
 netCDF4 = "^1.5.7"
 numba = "^0.57"
 typing-extensions = "^4.8.0"
-
-
-[tool.poetry.group.io]
-optional = false
-
-[tool.poetry.group.io.dependencies]
 zarr = ">=2.0.0"
 xarray-datatree = ">=0.0.5"
 
