### 2023-11-12 e48b3fd
2.0.3
Automatically generated by python-semantic-release
