### 2023-11-13 5879eb4
perf: compute scaler arrays simultaneously
