+++ xeofs/models/_base_model.py
@@ -65,9 +65,9 @@ class _BaseModel(ABC):
     feature_name: str, default="feature"
         Name of the feature dimension.
     compute: bool, default=True
-        Whether to compute the decomposition immediately. This is recommended
-        if the SVD result for the first ``n_modes`` can be accommodated in memory, as it
-        boosts computational efficiency compared to deferring the computation.
+        Whether to compute elements of the model eagerly, or to defer computation.
+        If True, the model's preprocessor scaler properties will be computed, followed
+        by the SVD decomposition, followed by the scores and components.
     verbose: bool, default=False
         Whether to show a progress bar when computing the decomposition.
     random_state: Optional[int], default=None
@@ -184,12 +184,12 @@ class _BaseModel(ABC):
             X, self.sample_dims, weights
         )
 
-        fitted_model = self._fit_algorithm(data2D)
+        self._fit_algorithm(data2D)
 
         if self._params["compute"]:
-            fitted_model.compute()
+            self.data.compute()
 
-        return fitted_model
+        return self
 
     @abstractmethod
     def _fit_algorithm(self, data: DataArray) -> Self:
