### 2023-11-13 6875a10
fix: passing solver kwargs
