### 2023-11-13 #6009 @samhforbes
Hi @kthyng I'm happy to take a run at this, as long as I can find a climate expert to review (and recommendations would be welcome here!)
