### 2023-11-13 55870e9
perf: compute nan checks as single task graph
