+++ xeofs/models/eof.py
@@ -30,10 +30,11 @@ class EOF(_BaseModel):
         Name of the sample dimension.
     feature_name: str, default="feature"
         Name of the feature dimension.
-    compute: bool, default=True
+    compute : bool, default=True
         Whether to compute elements of the model eagerly, or to defer computation.
-        If True, the model's preprocessor scaler properties will be computed, followed
-        by the SVD decomposition, followed by the scores and components.
+        If True, four pieces of the fit will be computed sequentially: 1) the
+        preprocessor scaler, 2) optional NaN checks, 3) SVD decomposition, 4) scores
+        and components.
     verbose: bool, default=False
         Whether to show a progress bar when computing the decomposition.
     random_state : Optional[int], default=None
@@ -275,10 +276,11 @@ class ComplexEOF(EOF):
         Name of the sample dimension.
     feature_name: str, default="feature"
         Name of the feature dimension.
-    compute: bool, default=True
+    compute : bool, default=True
         Whether to compute elements of the model eagerly, or to defer computation.
-        If True, the model's preprocessor scaler properties will be computed, followed
-        by the SVD decomposition, followed by the scores and components.
+        If True, four pieces of the fit will be computed sequentially: 1) the
+        preprocessor scaler, 2) optional NaN checks, 3) SVD decomposition, 4) scores
+        and components.
     verbose: bool, default=False
         Whether to show a progress bar when computing the decomposition.
     random_state : Optional[int], default=None
