### 2023-11-13 7071eeb
perf: avoid duplicate preprocessor computation
