### 2023-11-13 cd64310
feat: pass kwargs through compute to dask
