### 2023-11-14 c72f91c
Update docs/overview_2_features.rst
Co-authored-by: Niclas Rieger <45175997+nicrie@users.noreply.github.com>
