### 2023-11-14 63f21ca
Update xeofs/preprocessing/sanitizer.py
Co-authored-by: Niclas Rieger <45175997+nicrie@users.noreply.github.com>
