### 2023-11-14 8f91d95
Update xeofs/preprocessing/sanitizer.py
Co-authored-by: Niclas Rieger <45175997+nicrie@users.noreply.github.com>
