### 2023-11-14 5076088
Update xeofs/models/opa.py
Co-authored-by: Niclas Rieger <45175997+nicrie@users.noreply.github.com>
