### 2023-11-14 cfabbf3
perf: duplicative nan checks
