+++ xeofs/models/eeof.py
@@ -108,6 +108,7 @@ class ExtendedEOF(EOF):
                 standardize=False,
                 use_coslat=False,
                 compute=self._params["compute"],
+                check_nans=False,
                 sample_name=self.sample_name,
                 feature_name=self.feature_name,
                 solver_kwargs=self._params["solver_kwargs"],
@@ -145,6 +146,7 @@ class ExtendedEOF(EOF):
             standardize=False,
             use_coslat=False,
             compute=self._params["compute"],
+            check_nans=False,
             sample_name=self.sample_name,
             feature_name=self.feature_name,
             solver=self._params["solver"],
