### 2023-11-14 314901e
feat: lazy execution mode (#110)
