### 2023-11-14 #6009 @nicrie
Thank you for your time, @samhforbes ! I've checked the list of potential reviewers for climate/atmosphere experts:

- [rabernat](https://github.com/rabernat)
- [kgoebber](https://github.com/kgoebber)
- [DamienIrving](https://github.com/DamienIrving)
- [arbennett](https://github.com/arbennett)

For completeness, I will also list again the potential reviewers that I already mentioned above, who also seem to have a lot of experience in the relevant areas.

> * [malmans2](https://github.com/malmans2)
> * [miniufo](https://github.com/miniufo)
> * [syedhamidali](https://github.com/syedhamidali)
> * [jimmielin](https://github.com/jimmielin)
> * [eviatarbach](https://github.com/eviatarbach)
