### 2023-11-14 f2781d8
docs: acknowledge contributors in README (#115)
