### 2023-11-14 1a88379
docs: acknowledge contributors in README (#116)
