+++ README.rst
@@ -1,216 +1,0 @@
-.. image:: docs/logos/xeofs_logo.png
-  :align: center
-  :width: 800
-  :alt: xeofs logo
-
-
-+----------------------------+-----------------------------------------------------+
-| Versions                   | |pypi| |conda|                                      |
-+----------------------------+-----------------------------------------------------+
-| Build & Testing            | |build| |coverage|                                  |
-+----------------------------+-----------------------------------------------------+
-| Code Quality               | |black|                                             |
-+----------------------------+-----------------------------------------------------+
-| Documentation              | |docs|                                              |
-+----------------------------+-----------------------------------------------------+
-| Citation & Licensing       | |zenodo| |license|                                  |
-+----------------------------+-----------------------------------------------------+
-| User Engagement            | |downloads|                                         |
-+----------------------------+-----------------------------------------------------+
-
-.. |pypi| image:: https://img.shields.io/pypi/v/xeofs
-   :target: https://pypi.org/project/xeofs/
-   :alt: Python Package Index (PyPI)
-
-.. |conda| image:: https://img.shields.io/conda/vn/conda-forge/xeofs
-   :target: https://anaconda.org/conda-forge/xeofs
-   :alt: Conda-forge Version
-
-.. |build| image:: https://img.shields.io/github/actions/workflow/status/nicrie/xeofs/ci.yml?branch=main
-   :target: https://github.com/nicrie/xeofs/actions
-   :alt: Build Status
-
-.. |docs| image:: https://readthedocs.org/projects/xeofs/badge/?version=latest
-   :target: https://xeofs.readthedocs.io/en/latest/?badge=latest
-   :alt: Documentation Status
-
-.. |black| image:: https://img.shields.io/badge/code%20style-black-000000.svg
-   :target: https://github.com/psf/black
-   :alt: Black Code Style
-
-.. |coverage| image:: https://codecov.io/gh/nicrie/xeofs/branch/main/graph/badge.svg?token=8040ZDH6U7
-    :target: https://codecov.io/gh/nicrie/xeofs
-    :alt: Test Coverage
-
-.. |zenodo| image:: https://zenodo.org/badge/DOI/10.5281/zenodo.6323012.svg
-   :target: https://doi.org/10.5281/10.5281/zenodo.6323011
-   :alt: DOI - Zenodo
-
-.. |license| image:: https://img.shields.io/pypi/l/xeofs
-   :target: https://github.com/nicrie/xeofs/blob/main/LICENSE
-   :alt: License
-
-.. |downloads| image:: https://img.shields.io/pypi/dw/xeofs
-   :alt: PyPI - Downloads
-
-
-
-Overview
----------------------
-
-``xeofs`` is a dedicated Python package for dimensionality reduction in the realm of climate science, 
-offering methods like PCA, known as EOF analysis within the field, and related variants. 
-Seamlessly integrated with ``xarray`` and ``Dask``, it's tailored for easy handling and scalable 
-computation on large, multi-dimensional datasets, making advanced climate data analysis both accessible and efficient.
-
-- **Multi-Dimensional**: Designed for ``xarray`` objects, it applies dimensionality reduction to multi-dimensional data while maintaining data labels.
-- **Dask-Integrated**: Supports large datasets via ``Dask`` xarray objects
-- **Extensive Methods**: Offers various dimensionality reduction techniques
[…]
