### 2023-11-15 e015f47
refactor: set computed attrs via deserialize
