+++ xeofs/models/_base_model.py
@@ -349,7 +349,10 @@ class _BaseModel(ABC):
         **kwargs
             Additional keyword arguments to pass to `dask.compute()`.
         """
+        # find and compute all dask arrays simultaneously to allow dask to optimize the
+        # shared graph and avoid duplicate i/o and computations
         dt = self.serialize()
+
         data_objs = {
             k: v
             for k, v in dt.to_dict().items()
@@ -362,30 +365,11 @@ class _BaseModel(ABC):
         else:
             (data_objs,) = dask.compute(data_objs, **kwargs)
 
-        # This feels pretty fragile with all the casing, would be
-        # best to homogenize certain aspects of how we store data
-        # across different classes
-        for key, data in data_objs.items():
-            path_elems = key.strip("/").split("/")
-            parent = self
-            for elem in path_elems[:-1]:
-                if elem.isdigit():
-                    parent = parent[int(elem)]
-                else:
-                    parent = getattr(parent, elem)
+        for k, v in data_objs.items():
+            dt[k] = DataTree(v)
 
-            if isinstance(data, xr.Dataset):
-                mapping = data.attrs.get("name_map", {}).get(
-                    path_elems[-1], path_elems[-1]
-                )
-                da = data[mapping]
-            else:
-                da = data
-
-            if isinstance(parent, dict):
-                parent[path_elems[-1]] = da
-            else:
-                setattr(parent, path_elems[-1], da)
+        # then rebuild the trained model from the computed results
+        self._deserialize_attrs(dt)
 
         self._post_compute()
 
@@ -396,7 +380,7 @@ class _BaseModel(ABC):
         """Get the model parameters."""
         return self._params
 
-    def serialize(self, save_data: bool = False) -> DataTree:
+    def serialize(self) -> DataTree:
         """Serialize a complete model with its preprocessor."""
         # Create a root node for this object with its params as attrs
         ds_root = xr.Dataset(attrs=dict(params=self.get_params()))
@@ -405,7 +389,7 @@ class _BaseModel(ABC):
         # Retrieve the tree representation of each attached object, or set basic attrs
         for key, attr in self.get_serialization_attrs().items():
             if hasattr(attr, "serialize"):
-                dt[key] = attr.serialize(save_data=save_data)
+                dt[key] = attr.serialize()
                 dt.attrs[key] = "_is_tree"
             else:
                 dt.attrs[key] = attr
@@ -434,7 +418,21 @@ class _BaseModel(ABC):
 
         """
         self.compute()
-        dt = self.serialize(save_data=save_data)
+
+        dt = self.serialize()
+
+        # Remove any raw data arrays at this stage
+        for node in dt.subtree:
+            if not node.attrs.get("allow_compute", True) and not save_data:
+                dt[node.path] = DataTree(
+                    xr.Dataset(
+                        data_vars={
+                            node.name: xr.DataArray(np.nan, attrs={"placeholder": True})
+                        },
+                        attrs={"allow_compute": False, "placeholder": True},
+                    )
+                )
+
         write_mode = "w" if overwrite else "w-"
         dt.to_zarr(path, mode=write_mode, **kwargs)
 
@@ -444,14 +442,17 @@ class _BaseModel(ABC):
         # Recreate the model with parameters set by root level attrs
         params = dt.attrs.pop("params")
         model = cls(**params)
+        model._deserialize_attrs(dt)
+        return model
+
+    def _deserialize_attrs(self, dt: DataTree):
+        """Set the necessary attributes of the model from a DataTree."""
         for key, attr in dt.attrs.items():
             if attr == "_is_tree":
-                deserialized_obj = getattr(model, key).deserialize(dt[key])
[…]
