### 2023-11-15 ca4292b
fix: pyproject readme
