### 2023-11-17 1878406
refactor: set computed attrs via deserialize (#117)
