### 2023-11-17 65e9f79
2.1.1
Automatically generated by python-semantic-release
