### 2023-11-18 06a2014
docs: reactivate CHANGELOG
