### 2023-11-18 26f8345
ci: add workflow dispatch to Github Action
