+++ .github/workflows/changelog.yml
@@ -2,6 +2,7 @@ name: Generate changelog
 on:
   release:
     types: [created, edited]
+  workflow_dispatch: # Allows you to run this workflow manually from the Actions tab
 
 jobs:
   generate-changelog:
