### 2023-11-19 e7d6d90
feat: normalized option in inverse_transform
