### 2023-11-19 fa58532
docs(changelog): use PAT instead of GITHUB_TOKEN
