+++ .github/workflows/changelog.yml
@@ -1,4 +1,4 @@
-name: Generate changelog
+name: Update CHANGELOG
 on:
   release:
     types: [created, edited]
@@ -14,7 +14,7 @@ jobs:
     - uses: BobAnkh/auto-generate-changelog@v1.2.5
       with:
         REPO_NAME: 'nicrie/xeofs'
-        ACCESS_TOKEN: ${{secrets.GITHUB_TOKEN}}
+        ACCESS_TOKEN: ${{ secrets.CHANGELOG_TOKEN }}
         PATH: 'CHANGELOG.md'
         COMMIT_MESSAGE: 'docs(CHANGELOG): update release notes'
         TYPE: 'feat:Feature,fix:Bug Fixes,docs:Documentation,refactor:Refactor,perf:Performance Improvements'
