### 2023-11-19 42b5c5e
docs(CHANGELOG): update release notes
