+++ CHANGELOG.md
@@ -1,10 +1,448 @@
 # CHANGELOG
 
-
-
-## v1.0.4 (2023-07-22)
+## [v2.1.2](https://github.com/nicrie/xeofs/releases/tag/v2.1.2) - 2023-11-18 15:13:08
 
 ### Fix
+
+* Update default power iterations for improved accuracy ([#122](https://github.com/nicrie/xeofs/issues/122)) ([`9045bc6`](https://github.com/nicrie/xeofs/commit/9045bc68f18b34df2ff88c64de9131d8d26e3e38))
+
+### Documentation
+
+* Reactivate CHANGELOG ([#121](https://github.com/nicrie/xeofs/issues/121)) ([`6a4370b`](https://github.com/nicrie/xeofs/commit/6a4370b0bdfa31a679ce97b1aa7f1ab56041dbe4))
+
+### Bug Fixes
+
+- general:
+  - update default power iterations for improved accuracy (#122) ([9045bc6](https://github.com/nicrie/xeofs/commit/9045bc68f18b34df2ff88c64de9131d8d26e3e38)) ([#122](https://github.com/nicrie/xeofs/pull/122))
+
+### Documentation
+
+- general:
+  - reactivate CHANGELOG (#121) ([6a4370b](https://github.com/nicrie/xeofs/commit/6a4370b0bdfa31a679ce97b1aa7f1ab56041dbe4)) ([#121](https://github.com/nicrie/xeofs/pull/121))
+
+## [v2.1.1](https://github.com/nicrie/xeofs/releases/tag/v2.1.1) - 2023-11-17 16:45:04
+
+### Fix
+
+* Only sort transform if model is already sorted ([#118](https://github.com/nicrie/xeofs/issues/118)) ([`dfe0001`](https://github.com/nicrie/xeofs/commit/dfe0001a18a2ebd2073831e2b6484ace86ea0088))
+
+### Documentation
+
+* Fix broken link ([`de49038`](https://github.com/nicrie/xeofs/commit/de4903863b7c9fd53f8bae39ff6505bcdfadeb6e))
+* Acknowledge contributors in README ([#116](https://github.com/nicrie/xeofs/issues/116)) ([`1a88379`](https://github.com/nicrie/xeofs/commit/1a8837915bf89b83c58de1cdff4fe826b0f6d43d))
+
+### Bug Fixes
+
+- general:
+  - only sort transform if model is already sorted (#118) ([dfe0001](https://github.com/nicrie/xeofs/commit/dfe0001a18a2ebd2073831e2b6484ace86ea0088)) ([#118](https://github.com/nicrie/xeofs/pull/118))
+
+### Documentation
+
+- general:
+  - fix broken link ([de49038](https://github.com/nicrie/xeofs/commit/de4903863b7c9fd53f8bae39ff6505bcdfadeb6e))
+  - acknowledge contributors in README (#116) ([1a88379](https://github.com/nicrie/xeofs/commit/1a8837915bf89b83c58de1cdff4fe826b0f6d43d)) ([#116](https://github.com/nicrie/xeofs/pull/116))
+
+### Refactor
+
+- general:
+  - set computed attrs via deserialize (#117) ([1878406](https://github.com/nicrie/xeofs/commit/1878406702963f099c8a0b3ba2db494d4ef0bab4)) ([#117](https://github.com/nicrie/xeofs/pull/117))
+
+## [v2.1.0](https://github.com/nicrie/xeofs/releases/tag/v2.1.0) - 2023-11-14 12:03:45
+
+### Feature
+
[…]
