### 2023-11-20 8d06153
feat: support saving to netcdf (#123)
