+++ tests/models/test_eof.py
@@ -480,7 +480,8 @@ def test_inverse_transform(dim, mock_data_array):
         (("lon", "lat")),
     ],
 )
-def test_save_load(dim, mock_data_array, tmp_path):
+@pytest.mark.parametrize("engine", ["netcdf4", "zarr"])
+def test_save_load(dim, mock_data_array, tmp_path, engine):
     """Test save/load methods in EOF class, ensuring that we can
     roundtrip the model and get the same results when transforming
     data."""
@@ -488,13 +489,13 @@ def test_save_load(dim, mock_data_array, tmp_path):
     original.fit(mock_data_array, dim)
 
     # Save the EOF model
-    original.save(tmp_path / "eof.zarr")
+    original.save(tmp_path / "eof", engine=engine)
 
     # Check that the EOF model has been saved
-    assert (tmp_path / "eof.zarr").exists()
+    assert (tmp_path / "eof").exists()
 
     # Recreate the model from saved file
-    loaded = EOF.load(tmp_path / "eof.zarr")
+    loaded = EOF.load(tmp_path / "eof", engine=engine)
 
     # Check that the params and DataContainer objects match
     assert original.get_params() == loaded.get_params()
