+++ tests/models/test_mca.py
@@ -393,7 +393,8 @@ def test_compute(mock_dask_data_array, dim, compute):
         (("lon", "lat")),
     ],
 )
-def test_save_load(dim, mock_data_array, tmp_path):
+@pytest.mark.parametrize("engine", ["netcdf4", "zarr"])
+def test_save_load(dim, mock_data_array, tmp_path, engine):
     """Test save/load methods in MCA class, ensuring that we can
     roundtrip the model and get the same results when transforming
     data."""
@@ -401,13 +402,13 @@ def test_save_load(dim, mock_data_array, tmp_path):
     original.fit(mock_data_array, mock_data_array, dim)
 
     # Save the EOF model
-    original.save(tmp_path / "mca.zarr")
+    original.save(tmp_path / "mca", engine=engine)
 
     # Check that the EOF model has been saved
-    assert (tmp_path / "mca.zarr").exists()
+    assert (tmp_path / "mca").exists()
 
     # Recreate the model from saved file
-    loaded = MCA.load(tmp_path / "mca.zarr")
+    loaded = MCA.load(tmp_path / "mca", engine=engine)
 
     # Check that the params and DataContainer objects match
     assert original.get_params() == loaded.get_params()
