### 2023-11-20 a9d6bdf
fix: test loaded model vs transform not scores
