+++ tests/models/test_mca.py
@@ -422,16 +422,12 @@ def test_save_load(dim, mock_data_array, tmp_path):
 
     # Test that the recreated model can be used to transform new data
     assert np.allclose(
-        original.scores(),
+        original.transform(mock_data_array, mock_data_array),
         loaded.transform(mock_data_array, mock_data_array),
-        rtol=1e-3,
-        atol=1e-3,
     )
 
     # The loaded model should also be able to inverse_transform new data
     assert np.allclose(
         original.inverse_transform(*original.scores()),
         loaded.inverse_transform(*loaded.scores()),
-        rtol=1e-3,
-        atol=1e-3,
     )
