+++ tests/models/test_eof_rotator.py
@@ -235,12 +235,11 @@ def test_save_load(dim, mock_data_array, tmp_path):
 
     # Test that the recreated model can be used to transform new data
     assert np.allclose(
-        original.scores(), loaded.transform(mock_data_array), rtol=1e-3, atol=1e-3
+        original.transform(mock_data_array), loaded.transform(mock_data_array)
     )
 
     # The loaded model should also be able to inverse_transform new data
     assert np.allclose(
         original.inverse_transform(original.scores()),
         loaded.inverse_transform(loaded.scores()),
-        rtol=1e-2,
     )
