### 2023-11-20 c7c79e4
ci: specify Python Semantic Release version v8.0.0
