+++ .github/workflows/release_package.yml
@@ -24,7 +24,7 @@ jobs:
       # This action uses Python Semantic Release v8
       - name: Python Semantic Release
         id: release
-        uses: python-semantic-release/python-semantic-release@v8
+        uses: python-semantic-release/python-semantic-release@v8.0.0
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
 
