### 2023-11-20 901982d
2.2.0
Automatically generated by python-semantic-release
