+++ CHANGELOG.md
@@ -1,448 +1,713 @@
 # CHANGELOG
 
-## [v2.1.2](https://github.com/nicrie/xeofs/releases/tag/v2.1.2) - 2023-11-18 15:13:08
+
+
+## v2.2.0 (2023-11-20)
+
+### Chore
+
+* chore: release v2.2.0 (#128) ([`ef40a8a`](https://github.com/nicrie/xeofs/commit/ef40a8afa32ca5896aee75ad0cabcb943c66dfbc))
+
+### Ci
+
+* ci: specify Python Semantic Release version v8.0.0 ([`c7c79e4`](https://github.com/nicrie/xeofs/commit/c7c79e448462d96d49cec2ab903363085e5ed9d4))
+
+* ci: add workflow dispatch to Github Action ([`26f8345`](https://github.com/nicrie/xeofs/commit/26f834516d1aacecdffc91e425a801c2086bdcd8))
+
+### Documentation
+
+* docs(CHANGELOG): update release notes ([`42b5c5e`](https://github.com/nicrie/xeofs/commit/42b5c5ebb1e98c20a98c3ab770eb94000243b6d2))
+
+* docs(changelog): use PAT instead of GITHUB_TOKEN ([`fa58532`](https://github.com/nicrie/xeofs/commit/fa5853203b514146b295d9185e436295fed18c81))
+
+### Feature
+
+* feat: normalized option in inverse_transform (#124) ([`4ad2791`](https://github.com/nicrie/xeofs/commit/4ad2791d4db672c7c722ee2e5d12cf3b3a88028c))
+
+* feat: support saving to netcdf (#123) ([`8d06153`](https://github.com/nicrie/xeofs/commit/8d06153b5de31aad3f0817979c1122250f444c57))
+
+* feat: normalized option in inverse_transform ([`e7d6d90`](https://github.com/nicrie/xeofs/commit/e7d6d90d6fdcbd4575cddf3b12ebf1852dc329ef))
 
 ### Fix
 
-* Update default power iterations for improved accuracy ([#122](https://github.com/nicrie/xeofs/issues/122)) ([`9045bc6`](https://github.com/nicrie/xeofs/commit/9045bc68f18b34df2ff88c64de9131d8d26e3e38))
+* fix: update to handle v8 semantic release GH action (#127) ([`5b8c6d8`](https://github.com/nicrie/xeofs/commit/5b8c6d83f10d9439b0ced8d72c06c1fb36804fe4))
+
+* fix: test loaded model vs transform not scores ([`a9d6bdf`](https://github.com/nicrie/xeofs/commit/a9d6bdf21278477bb42d7a406aed6053dee9c223))
+
+### Unknown
+
+* Merge branch &#39;main&#39; into develop ([`af25e8c`](https://github.com/nicrie/xeofs/commit/af25e8cf989d5a68fa362a72062f211b9a392798))
+
+
+## v2.1.2 (2023-11-18)
+
+### Chore
+
+* chore: Merge branch &#39;main&#39; into develop ([`607d25b`](https://github.com/nicrie/xeofs/commit/607d25bd827f849dbd40c5fb0b68ca93a95e45a5))
 
 ### Documentation
 
-* Reactivate CHANGELOG ([#121](https://github.com/nicrie/xeofs/issues/121)) ([`6a4370b`](https://github.com/nicrie/xeofs/commit/6a4370b0bdfa31a679ce97b1aa7f1ab56041dbe4))
+* docs: reactivate CHANGELOG (#121) ([`6a4370b`](https://github.com/nicrie/xeofs/commit/6a4370b0bdfa31a679ce97b1aa7f1ab56041dbe4))
 
-### Bug Fixes
-
-- general:
[…]
