### 2023-11-20 aba57e0
2.2.1
Automatically generated by python-semantic-release
