### 2023-11-20 320dbcb
chore(release): Updated semantic-release configuration
