+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 
 
+## v2.2.2 (2023-11-20)
+
+### Chore
+
+* chore(release): update PSR version ([`0f5524f`](https://github.com/nicrie/xeofs/commit/0f5524fddafbd5aa71df5de40006f5ace75561bf))
+
+* chore(release): Updated semantic-release
+configuration ([`320dbcb`](https://github.com/nicrie/xeofs/commit/320dbcbb252340343fa6eaef794a8941c2fe03de))
+
+### Fix
+
+* fix: trigger new release ([`c79a116`](https://github.com/nicrie/xeofs/commit/c79a1160d5b27e44f82b9c8fe9a9ac808ffca895))
+
+
 ## v2.2.1 (2023-11-20)
 
 ### Ci
@@ -11,6 +25,12 @@
 ### Fix
 
 * fix: trigger new release ([`58250b9`](https://github.com/nicrie/xeofs/commit/58250b948bf1d2316a589b3872a44d3d7ecb7ec9))
+
+### Unknown
+
+* 2.2.1
+
+Automatically generated by python-semantic-release ([`aba57e0`](https://github.com/nicrie/xeofs/commit/aba57e0d97c792ffd5449486bb6cd10b6e534a86))
 
 
 ## v2.2.0 (2023-11-20)
@@ -47,6 +67,10 @@
 
 ### Unknown
 
+* 2.2.0
+
+Automatically generated by python-semantic-release ([`901982d`](https://github.com/nicrie/xeofs/commit/901982d6df1c73b21bd222cc54f7a6e50a683d56))
+
 * Merge branch &#39;main&#39; into develop ([`af25e8c`](https://github.com/nicrie/xeofs/commit/af25e8cf989d5a68fa362a72062f211b9a392798))
 
 
@@ -65,6 +89,12 @@
 ### Fix
 
 * fix: update default power iterations for improved accuracy (#122) ([`9045bc6`](https://github.com/nicrie/xeofs/commit/9045bc68f18b34df2ff88c64de9131d8d26e3e38))
+
+### Unknown
+
+* 2.1.2
+
+Automatically generated by python-semantic-release ([`d80a839`](https://github.com/nicrie/xeofs/commit/d80a839dc0c301557e7ec58559197a5f40c7353d))
 
 
 ## v2.1.1 (2023-11-17)
@@ -97,6 +127,12 @@ Implement solution suggested in all-contributors/app#485 ([`8929ed5`](https://gi
 
 * refactor: set computed attrs via deserialize (#117) ([`1878406`](https://github.com/nicrie/xeofs/commit/1878406702963f099c8a0b3ba2db494d4ef0bab4))
 
+### Unknown
+
+* 2.1.1
+
+Automatically generated by python-semantic-release ([`65e9f79`](https://github.com/nicrie/xeofs/commit/65e9f79a880fc2ca24b60961e5271ccbad1a2084))
+
 
 ## v2.1.0 (2023-11-14)
 
@@ -114,6 +150,12 @@ Implement solution suggested in all-contributors/app#485 ([`8929ed5`](https://gi
 
 * feat: lazy execution mode (#110) ([`314901e`](https://github.com/nicrie/xeofs/commit/314901ebd978290d56214219d7a87f5111f07775))
 
+### Unknown
+
+* 2.1.0
+
+Automatically generated by python-semantic-release ([`a3505c9`](https://github.com/nicrie/xeofs/commit/a3505c99abcc8329033d221016b2d2b85e2528bf))
+
 
 ## v2.0.3 (2023-11-12)
 
@@ -125,6 +167,12 @@ Implement solution suggested in all-contributors/app#485 ([`8929ed5`](https://gi
 
[…]
