### 2023-12-13 #6060 @DamienIrving
I'm all done - sorry for the delay!

The authors have done a really excellent job with the package and associated documentation.

The only very minor issue I found was a broken link to the contributing guide, for which I've submitted a pull request to fix: https://github.com/nicrie/xeofs/pull/129
