### 2023-12-14 #6060 @malmans2
Hi all, sorry about the delay!

I glanced through the repository and `xeofs` looks in great shape.
I'll go through all the formal steps by the end of the week and will let you know when I'm done.
