### 2023-12-14 38ba677
Fix link to perf scripts
