### 2023-12-14 nicrie/xeofs#130 @malmans2
Keep the documentation in sync with the code
Hi there,

Great package and awesome documentation!

I have a small suggestion.
I often find hard to keep the documentation in sync with the code. I suggest to store the simple notebooks with no output (e.g., you could add [nbstripout](https://github.com/kynan/nbstripout) to your pre-commits) and make readthedocs executing them.
I imagine the example notebooks are too heavy to do that, so I'm referring to the simple notebooks such as [quickstart](https://github.com/nicrie/xeofs/blob/main/docs/quickstart.ipynb).

Similarly, I would use doctests in the README rather than simple snippets. If you go that way, you just need to tell pytest to check the doctests in README.md, and you are sure that the README is in sync with the code.
See: https://docs.pytest.org/en/7.3.x/how-to/doctest.html#how-to-run-doctests

JOSS review: https://github.com/openjournals/joss-reviews/issues/6060
