### 2023-12-15 nicrie/xeofs#132 @malmans2
Add dependencies section in the documentation
I suggest to add a dependencies section in the documentation.
`xarray`'s documentation is a good example: https://docs.xarray.dev/en/stable/getting-started-guide/installing.html#installation

JOSS review: https://github.com/openjournals/joss-reviews/issues/6060
