### 2023-12-15 #6060 @malmans2
Hi everyone! I'm done with my review.

Great package, and it's 100% in line with JOSS standards.
If I'll need to perform EOFs, I will definitely use it.

I opened 2 minor issues:
- https://github.com/nicrie/xeofs/issues/132: Although the authors clearly mention the core dependencies, I think it's worth listing somewhere in the documentation all mandatory and optional dependencies.
- https://github.com/nicrie/xeofs/issues/130: This shouldn't be a blocker, so it's up to the authors to decide whether implement my suggestions.

I suggest to open a PR in xarray and add `xeofs` here: https://docs.xarray.dev/en/stable/ecosystem.html
