### 2023-12-16 95f29aa
docs: use doctest to test README examples
