### 2023-12-17 f36e404
build: loose version constraints of dependencies
