+++ pyproject.toml
@@ -12,18 +12,18 @@ documentation = "https://xeofs.readthedocs.io/en/latest/"
 [tool.poetry.dependencies]
 python = ">=3.10,<3.12"
 poetry = "^1.6"
-numpy = "~1.24"
-pandas = ">=1.4.1"
-xarray = ">=0.21.1"
-scikit-learn = ">=1.0.2"
+numpy = "^1.24"
+pandas = "^2"
+xarray = ">=2023"
+scikit-learn = "^1.0.2"
 pooch = "^1.6.0"
 tqdm = "^4.64.0"
 dask = ">=2023.0.1"
-statsmodels = ">=0.14.0"
+statsmodels = "^0.14.0"
 netCDF4 = "^1.5.7"
-numba = "^0.57"
+numba = ">=0.57"
 typing-extensions = "^4.8.0"
-zarr = ">=2.0.0"
+zarr = "^2.0.0"
 xarray-datatree = ">=0.0.5"
 
 [tool.poetry.group.dev]
@@ -40,7 +40,7 @@ pre-commit = "^3.3"
 optional = true
 
 [tool.poetry.group.docs.dependencies]
-rpy2 = ">=3.5"
+rpy2 = "^3.5"
 sphinx-gallery = "^0.14"
 sphinx-design = "^0.5"
 sphinx-copybutton = "^0.5"
