### 2023-12-17 nicrie/xeofs#132 @nicrie
Thank you @malmans2  for the suggestion, that makes sense (underlined by the recent question in #134 ;) ) ! I just made a PR to address this.
