### 2023-12-17 nicrie/xeofs#130 @slevang
Looks like the whole serialize/deserialize pathway doesn't quite work with Datasets. That's an issue since `.compute()` now relies on temporarily converting to a DataTree to compute everything at once. I'll try to look at getting this properly supported in the next few days. In the meantime your example works fine on the DataArray:
```python
t2m = xr.tutorial.open_dataset("air_temperature")
# fails
xe.models.EOF(n_modes=10).fit(t2m, "time").compute()
# succeeds
xe.models.EOF(n_modes=10).fit(t2m.air, "time").compute()
```
