### 2023-12-17 ff9a1be
fix: test serialization roundtrip with dataset
