### 2023-12-18 275dd45
docs: add spaces after code blocks for doctest
