### 2023-12-18 382f6c7
docs: use doctest to test README examples (#135)
