### 2023-12-19 #6060 @nicrie
Hi, many thanks @DamienIrving and @malmans2 for the reviews and the suggestions for improvement. We have incorporated the feedback and opened a PR for the xarray documentation [here](https://github.com/pydata/xarray/pull/8561).
