### 2023-12-20 #6060 @malmans2
> Can you confirm you are happy with the changes made and there are no outstanding concerns please?

Correct! In my opinion `xeofs` is ready for publication.
