### 2023-12-21 8035db9
docs: add JOSS badge
