### 2023-12-21 609386e
docs: update author list and homepage
