+++ pyproject.toml
@@ -2,7 +2,7 @@
 name = "xeofs"
 version = "2.2.3"
 description = "Comprehensive EOF analysis in Python with xarray: A versatile, multidimensional, and scalable tool for advanced climate data analysis"
-authors = ["Niclas Rieger <niclasrieger@gmail.com>", Samuel J. Levang]
+authors = ["Niclas Rieger <niclasrieger@gmail.com>", "Samuel J. Levang"]
 license = "MIT"
 readme = "README.md"
 homepage = "https://github.com/xarray-contrib/xeofs"
