+++ docs/overview_2_features.rst
@@ -193,7 +193,7 @@ significant speed advantages. The dashed line marks data sets with about 3 MiB;
     You can find the script to run the performance tests here_.
 
 
-.. _here: https://github.com/nicrie/xeofs/tree/main/docs/perf
+.. _here: https://github.com/xarray-contrib/xeofs/tree/main/docs/perf
 
 Implement Your Own Model
 -------------------------
@@ -392,5 +392,5 @@ toolkit for dimensionality reduction techniques.
 .. _xMCA: https://github.com/Yefee/xMCA
 .. _eofs: https://github.com/ajdawson/eofs
 .. _`sklearn documentation on PCA`: https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.PCA.html
-.. _`GitHub`: https://github.com/nicrie/xeofs/issues
+.. _`GitHub`: https://github.com/xarray-contrib/xeofs/issues
 
