+++ .github/workflows/changelog.yml
@@ -13,7 +13,7 @@ jobs:
         fetch-depth: 0
     - uses: BobAnkh/auto-generate-changelog@v1.2.5
       with:
-        REPO_NAME: 'nicrie/xeofs'
+        REPO_NAME: 'xarray-contrib/xeofs'
         ACCESS_TOKEN: ${{ secrets.CHANGELOG_TOKEN }}
         PATH: 'CHANGELOG.md'
         COMMIT_MESSAGE: 'docs(CHANGELOG): update release notes'
