### 2023-12-23 c3e4780
fix: bump new release for JOSS (#143)
Co-authored-by: Sam Levang <slevang@salientpredictions.com>
