+++ CHANGELOG.md
@@ -6,231 +6,231 @@
 
 ### Documentation
 
-* docs: show contributors in readme (#140) ([`0c12fa5`](https://github.com/nicrie/xeofs/commit/0c12fa5385e807e2cd45599e6fc0f10b3a520119))
+* docs: show contributors in readme (#140) ([`0c12fa5`](https://github.com/xarray-contrib/xeofs/commit/0c12fa5385e807e2cd45599e6fc0f10b3a520119))
 
-* docs: remove all-contributor config file ([`f5234fc`](https://github.com/nicrie/xeofs/commit/f5234fc3895f3bcd8893ae684ea23a0ebe001db4))
+* docs: remove all-contributor config file ([`f5234fc`](https://github.com/xarray-contrib/xeofs/commit/f5234fc3895f3bcd8893ae684ea23a0ebe001db4))
 
-* docs: add required dependencies (#138) ([`7ec694f`](https://github.com/nicrie/xeofs/commit/7ec694f41d6aca0509782096cbabbcd19264180a))
+* docs: add required dependencies (#138) ([`7ec694f`](https://github.com/xarray-contrib/xeofs/commit/7ec694f41d6aca0509782096cbabbcd19264180a))
 
-* docs: add required dependencies ([`ce6b93a`](https://github.com/nicrie/xeofs/commit/ce6b93ab31da6ab10ebc70e9444f1b9956ec02b2))
+* docs: add required dependencies ([`ce6b93a`](https://github.com/xarray-contrib/xeofs/commit/ce6b93ab31da6ab10ebc70e9444f1b9956ec02b2))
 
-* docs: fix link to perf scripts (#131) ([`e2b3276`](https://github.com/nicrie/xeofs/commit/e2b32762bd6bb3111c973bbd571b56367a496eba))
+* docs: fix link to perf scripts (#131) ([`e2b3276`](https://github.com/xarray-contrib/xeofs/commit/e2b32762bd6bb3111c973bbd571b56367a496eba))
 
-* docs: Update link to contributing guide (#129) ([`21b8d3d`](https://github.com/nicrie/xeofs/commit/21b8d3dd3cf46fb631f7c2fe28fdf95246cd4153))
+* docs: Update link to contributing guide (#129) ([`21b8d3d`](https://github.com/xarray-contrib/xeofs/commit/21b8d3dd3cf46fb631f7c2fe28fdf95246cd4153))
 
-* docs: show contributors in README ([`049e654`](https://github.com/nicrie/xeofs/commit/049e6544637d4c396d40ffcb1b29b7a710c584de))
+* docs: show contributors in README ([`049e654`](https://github.com/xarray-contrib/xeofs/commit/049e6544637d4c396d40ffcb1b29b7a710c584de))
 
 ### Fix
 
-* fix: serialization for dataset (#139) ([`93fce38`](https://github.com/nicrie/xeofs/commit/93fce38ca170b1c6c9adb32b345a4531f2db6fdf))
+* fix: serialization for dataset (#139) ([`93fce38`](https://github.com/xarray-contrib/xeofs/commit/93fce38ca170b1c6c9adb32b345a4531f2db6fdf))
 
-* fix: test serialization roundtrip with dataset ([`ff9a1be`](https://github.com/nicrie/xeofs/commit/ff9a1bea1441e0c742ca049750352b110b6376b1))
+* fix: test serialization roundtrip with dataset ([`ff9a1be`](https://github.com/xarray-contrib/xeofs/commit/ff9a1bea1441e0c742ca049750352b110b6376b1))
 
[…]
