+++ pyproject.toml
@@ -2,11 +2,11 @@
 name = "xeofs"
 version = "2.2.3"
 description = "Comprehensive EOF analysis in Python with xarray: A versatile, multidimensional, and scalable tool for advanced climate data analysis"
-authors = ["Niclas Rieger <niclasrieger@gmail.com>"]
+authors = ["Niclas Rieger <niclasrieger@gmail.com>", "Samuel J. Levang <slevang@gmail.com>"]
 license = "MIT"
 readme = "README.md"
-homepage = "https://github.com/nicrie/xeofs"
-repository = "https://github.com/nicrie/xeofs"
+homepage = "https://github.com/xarray-contrib/xeofs"
+repository = "https://github.com/xarray-contrib/xeofs"
 documentation = "https://xeofs.readthedocs.io/en/latest/"
 
 [tool.poetry.dependencies]
