+++ CHANGELOG.md
@@ -2,7 +2,30 @@
 
 
 
+## v2.2.4 (2023-12-23)
+
+### Build
+
+* build: loose version constraints of dependencies (#137) ([`1a0efe1`](https://github.com/xarray-contrib/xeofs/commit/1a0efe1cbd18d5ce444a6e42eb2f290c9619dcb0))
+
+### Documentation
+
+* docs: automatically sync quickstart tutorial with codebase (#136) ([`48e63a0`](https://github.com/xarray-contrib/xeofs/commit/48e63a09579ff559908183ea2f13032886d64290))
+
+* docs: use doctest to test README examples (#135) ([`382f6c7`](https://github.com/xarray-contrib/xeofs/commit/382f6c7db5496d6a8921e3044add4acc8acb75d5))
+
+### Fix
+
+* fix: bump new release for JOSS (#143)
+
+Co-authored-by: Sam Levang &lt;slevang@salientpredictions.com&gt; ([`c3e4780`](https://github.com/xarray-contrib/xeofs/commit/c3e4780e915c2c72ecdf21739742b5b8a5ad07bd))
+
+
 ## v2.2.3 (2023-12-18)
+
+### Chore
+
+* chore(release): v2.2.3 ([`b149477`](https://github.com/xarray-contrib/xeofs/commit/b1494771cddab30812fcdb3633f42c871034ac89))
 
 ### Documentation
 
@@ -901,11 +924,11 @@ Automatically generated by python-semantic-release ([`75158c6`](https://github.c
 
 * chore: merge main into develop ([`9f50392`](https://github.com/xarray-contrib/xeofs/commit/9f50392d7f99ca6bb7cd13608b630094e0845cb3))
 
-* chore: merge branch &#39;main&#39; of xarray-contrib/xeofs ([`efb5ab9`](https://github.com/xarray-contrib/xeofs/commit/efb5ab9c306e9d71a4aab20b6f16ccbfa6a65c56))
+* chore: merge branch &#39;main&#39; of nicrie/xeofs ([`efb5ab9`](https://github.com/xarray-contrib/xeofs/commit/efb5ab9c306e9d71a4aab20b6f16ccbfa6a65c56))
 
 * chore: Merge branch &#39;main&#39; into develop ([`bc4e87d`](https://github.com/xarray-contrib/xeofs/commit/bc4e87d469807c4b2bb4679556813e2693affe5f))
 
-* chore: Merge  develop of xarray-contrib/xeofs into develop ([`f31e56e`](https://github.com/xarray-contrib/xeofs/commit/f31e56e49c550f87c6b132951f97ebd561ec5db5))
+* chore: Merge  develop of nicrie/xeofs into develop ([`f31e56e`](https://github.com/xarray-contrib/xeofs/commit/f31e56e49c550f87c6b132951f97ebd561ec5db5))
 
 ### Documentation
 
@@ -947,7 +970,7 @@ Automatically generated by python-semantic-release ([`75158c6`](https://github.c
 
 ### Refactor
 
-* refactor: Merge &#39;main&#39; of github.com:xarray-contrib/xeofs ([`93fb010`](https://github.com/xarray-contrib/xeofs/commit/93fb010103923d839295fce67cb012e0b70d157f))
+* refactor: Merge &#39;main&#39; of github.com:nicrie/xeofs ([`93fb010`](https://github.com/xarray-contrib/xeofs/commit/93fb010103923d839295fce67cb012e0b70d157f))
 
 ### Unknown
 
@@ -976,7 +999,7 @@ Automatically generated by python-semantic-release ([`464d71e`](https://github.c
 
[…]
