### 2024-01-02 d12dcc5
docs: replace hyphen with dash
