### 2024-01-02 370d3c7
docs: ensure capitalization of reference titles
