+++ publication/paper.bib
@@ -1,6 +1,6 @@
 @misc{xmca_yefee,
-	author = {He, Chengfei},
-  	title = {xMCA: Maximum Covariance Analysis in xarray for Climate Science},
+	author = {{He}, {Chengfei}},
+  	title = {{xMCA: Maximum Covariance Analysis in xarray for Climate Science}},
   	year = {2019},
   	publisher = {GitHub},
   	journal = {GitHub repository},
@@ -8,7 +8,7 @@
 }
 
 @article{dawson_eofs_2016,
-	title = {eofs: A Library for {EOF} Analysis of Meteorological, Oceanographic, and Climate Data},
+	title = {{eofs: A Library for {EOF} Analysis of Meteorological, Oceanographic, and Climate Data}},
 	volume = {4},
 	issn = {2049-9647},
 	url = {https://openresearchsoftware.metajnl.com/articles/10.5334/jors.122},
@@ -17,17 +17,16 @@
 	abstract = {The eofs library provides a high-level Python interface for computing empirical orthogonal functions ({EOFs}) and related quantities, with a focus on correctness and ease of use. The library is implemented in a modular hierarchical fashion, allowing computations using plain arrays, or the inclusion of metadata. The software provides a convenient package for users wanting to perform {EOF} analysis in Python, and integrates with popular libraries from atmospheric and climate science. The software is available on Github.},
 	pages = {e14},
 	number = {1},
-	author = {Dawson, Andrew},
+	author = {{Dawson}, {Andrew}},
 	urldate = {2023-10-30},
 	date = {2016-04-26},
 	langid = {american},
 	note = {Number: 1
 Publisher: Ubiquity Press},
-	file = {Full Text PDF:/home/nrieger/Zotero/storage/RDYCWBQ4/Dawson - 2016 - eofs A Library for EOF Analysis of Meteorological.pdf:application/pdf},
 }
 
 @Manual{dask2016,
-  title = {Dask: Library for dynamic task scheduling},
+  title = {{Dask: Library for dynamic task scheduling}},
   author = {{Dask Development Team}},
   year = {2016},
   url = {https://dask.org},
@@ -35,7 +34,7 @@ Publisher: Ubiquity Press},
 
 @article{hoyer2017xarray,
   title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
-  author    = {Hoyer, S. and J. Hamman},
+  author    = {{Hoyer}, S. and J. {Hamman}},
   journal   = {Journal of Open Research Software},
   volume    = {5},
   number    = {1},
@@ -48,7 +47,7 @@ Publisher: Ubiquity Press},
 
 @incollection{golub_singular_1971,
 	location = {Berlin, Heidelberg},
-	title = {Singular Value Decomposition and Least Squares Solutions},
+	title = {{Singular Value Decomposition and Least Squares Solutions}},
 	isbn = {978-3-642-86940-2},
 	url = {https://doi.org/10.1007/978-3-642-86940-2_10},
 	series = {Die Grundlehren der mathematischen Wissenschaften},
@@ -56,7 +55,7 @@ Publisher: Ubiquity Press},
 	pages = {134--151},
 	booktitle = {Handbook for Automatic Computation: Volume {II}: Linear Algebra},
 	publisher = {Springer},
-	author = {Golub, G. H. and Reinsch, C.},
+	author = {{Golub}, G. H. and {Reinsch}, C.},
[…]
