### 2024-01-02 2e28ae7
docs: remove parantheses around references in table
