### 2024-01-02 #6060 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
